+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased][]
 
+## [1.3.0][] - 2024-11-21
+
 ### Changed
 
 -   Pin `domain-browser` to 4.22.0
@@ -49,5 +51,6 @@
 [1.0.0]: https://github.com/niksy/node-stdlib-browser/tree/v1.0.0
 [1.1.0]: https://github.com/niksy/node-stdlib-browser/tree/v1.1.0
 [1.2.0]: https://github.com/niksy/node-stdlib-browser/tree/v1.2.0
-[Unreleased]: https://github.com/niksy/node-stdlib-browser/compare/v1.2.1...HEAD
+[Unreleased]: https://github.com/niksy/node-stdlib-browser/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/niksy/node-stdlib-browser/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/niksy/node-stdlib-browser/tree/v1.2.1